requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
djangorestframework-simplejwt
flake8
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
Pillow>=8.2.0,<8.3.0
uwsgi>=2.0.19,<2.1
django-cors-headers==3.13.0
certbot
django-filter==2.4.0
pandas==1.3.*
django-safedelete
